todo update
about 1000 images should be ok for now; the more images there are, the slower the system gets. I tested a gallery with 200 000 images, which took about 10 seconds of execution time, because there are a lot of calls for every image. so, technically it works with thousands of images, but is very slow. this could probably be enhanced by some PHP caching on the server, I didn't play around with it yet.

this state is good enough for now. if I need a gallery with multiple of thousand of images, I probably would build some static site generator thing, that generates a static HTML page and all image sizes beforehand.