make sure to add some event handlers only once
the Ajax.navigate() function re-inits all event listeners, because most of the content changed and we need to attach new events to the fullscreen button, the navigation buttons and so on. but we also add some event listeners directly to the dom or other elements that do not get replaced, so we need to make sure to not re-add those event listeners as well.