todo update
the Folder() class doesn't need to read the cache folder; only if we clean it out.
if there was an .placeholder subfolder, it would not be seen for the cleaning, and thus the 'delete empty folders' section of the refresh_cache() function wouldn't work, because there still would be that hidden .placeholder folder.

with a .placeholder suffix though, the files get cleaned out automatically. it should have no performance impact in having them directly in the cache folder instead of a subfolder.

thus, this todo is rendered moot.